backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
azure-search-documents>=11.4.0
azure-identity>=1.14.0
openai>=1.3.0
python-dotenv>=1.0.0
httpx>=0.25.0